move message-service up a level add smoke test for message-service
