+++ test/integration/message-service.test.js
@@ -1,0 +1,11 @@
+
+const messageService = require('../../server/services/message-service')
+
+describe('message service', () => {
+  afterEach(async () => {
+    await messageService.closeConnections()
+  })
+  test('smoke test', async () => {
+    await messageService.registerQueues()
+  })
+})
